requirements.txt
websockets==10.0
python-binance==1.0.16
sqlalchemy==1.4.15
schedule==1.1.0
apprise==0.9.5.1
Flask==2.0.1
gunicorn==20.1.0
eventlet==0.30.2
flask-cors==3.0.10
flask-socketio==5.0.1
python-socketio[client]==5.2.1
cachetools==4.2.2
sqlitedict==1.7.0
unicorn-binance-websocket-api==1.34.2
unicorn-fy==0.11.0
sortedcontainers
tabulate
prettytable
pandas
tqdm